Support gRPC Headers from ClientOptions and correct Scope Handling (#1039)
Allow additional headers to be added via ClientOptions, similar to TChannel. Additionally support removing headers by setting built-in headers values to null or an empty string.

Refactor OpenTracingInterceptor to a separate file and correct Scope handling to resolve Context mismatches. Leaving the Scope open during initialization and then closing it during onClose is incorrect and can cause issues with other interceptors. It's difficult to test this change in isolation.